deep_dive_sales_assistant/pyproject.toml
[tool.poetry]
name = "deep-dive-sales-assistant"
version = "0.1.0"
description = "AI-powered sales intelligence tool for company research and prospect analysis"
authors = ["Sales Team <sales@company.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
typer = {extras = ["all"], version = "^0.9.0"}
httpx = {extras = ["http2"], version = "^0.25.0"}
python-dotenv = "^1.0.0"
openai = "^1.14.0"
tldextract = "^5.1.0"
aioretry = "^6.3.0"
rapidfuzz = "^3.5.0"
toolz = "^0.12.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
respx = "^0.20.0"

[tool.poetry.scripts]
deepdive = "prospect:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"] 